major-projects/skin-cancer-cnn/requirements.txt
flask
tensorflow
numpy
pandas
pillow
scikit-learn
opencv-python-headless
werkzeug
matplotlib
seaborn
